nvif: split out perfmon interface definitions
